refactor(rfhcarousel): refactor out RfhCarousel as a separate component
