+++ src/components/RfhCarousel/index.ts
@@ -1,0 +1,1 @@
+export { default as RfhCarousel } from './RfhCarousel'
